extract RunningPhaseDeploymentStateResolver and implementations from KubernetesAppInstanceStatus
