+++ src/main/java/org/springframework/cloud/deployer/spi/kubernetes/CompositeDeploymentStateResolver.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *  
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package org.springframework.cloud.deployer.spi.kubernetes;
+
+/**
+ * @author David Turanski
+ **/
+
+import io.fabric8.kubernetes.api.model.ContainerStatus;
+import org.springframework.cloud.deployer.spi.app.DeploymentState;
+
+class CompositeDeploymentStateResolver implements RunningPhaseDeploymentStateResolver {
+	private final RunningPhaseDeploymentStateResolver[] delegates;
+
+	CompositeDeploymentStateResolver(RunningPhaseDeploymentStateResolver... delegates) {
+		this.delegates = delegates;
+	}
+
+	@Override
+	public DeploymentState resolve(ContainerStatus containerStatus) {
+		for (RunningPhaseDeploymentStateResolver resolver: delegates) {
+			DeploymentState deploymentState = resolver.resolve(containerStatus);
+			if (deploymentState != null) {
+				return deploymentState;
+			}
+		}
+		return null;
+	}
+}
