+++ src/main/java/org/springframework/cloud/deployer/spi/kubernetes/DefaultRunningPhaseDeploymentStateResolver.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package org.springframework.cloud.deployer.spi.kubernetes;
+
+import org.springframework.cloud.deployer.spi.app.DeploymentState;
+
+/**
+ * @author David Turanski
+ **/
+public class DefaultRunningPhaseDeploymentStateResolver extends CompositeDeploymentStateResolver {
+
+	public DefaultRunningPhaseDeploymentStateResolver(KubernetesDeployerProperties properties) {
+		super(
+			new PredicateRunningPhaseDeploymentStateResolver.ContainerReady(properties),
+			new PredicateRunningPhaseDeploymentStateResolver.ContainerCrashed(properties),
+			new PredicateRunningPhaseDeploymentStateResolver.RestartsDueToTheSameError(properties),
+			new PredicateRunningPhaseDeploymentStateResolver.CrashLoopBackOffRestarts(properties),
+			new PredicateRunningPhaseDeploymentStateResolver.ContainerTerminated(properties),
+			//default
+			containerStatus -> DeploymentState.deploying);
+	}
+}
